# backend/api/gen/go/app/service/v1/i_product_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: app/service/v1/i_product.proto

package servicev1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-shop/api/gen/go/catalog/service/v1"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationProductServiceGet = "/app.service.v1.ProductService/Get"
const OperationProductServiceList = "/app.service.v1.ProductService/List"

type ProductServiceHTTPServer interface {
	// Get 查询商品详情
	Get(context.Context, *v11.GetProductRequest) (*v11.Product, error)
	// List 分页查询商品列表
	List(context.Context, *v1.PagingRequest) (*v11.ListProductResponse, error)
}

func RegisterProductServiceHTTPServer(s *http.Server, srv ProductServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/app/v1/mall/products", _ProductService_List9_HTTP_Handler(srv))
	r.GET("/app/v1/mall/products/{id}", _ProductService_Get9_HTTP_Handler(srv))
}

func _ProductService_List9_HTTP_Handler(srv ProductServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListProductResponse)
		return ctx.Result(200, reply)
	}
}

func _ProductService_Get9_HTTP_Handler(srv ProductServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetProductRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationProductServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetProductRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Product)
		return ctx.Result(200, reply)
	}
}

type ProductServiceHTTPClient interface {
	// Get 查询商品详情
	Get(ctx context.Context, req *v11.GetProductRequest, opts ...http.CallOption) (rsp *v11.Product, err error)
	// List 分页查询商品列表
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListProductResponse, err error)
}

type ProductServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewProductServiceHTTPClient(client *http.Client) ProductServiceHTTPClient {
	return &ProductServiceHTTPClientImpl{client}
}

// Get 查询商品详情
func (c *ProductServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetProductRequest, opts ...http.CallOption) (*v11.Product, error) {
	var out v11.Product
	pattern := "/app/v1/mall/products/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationProductServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List 分页查询商品列表
func (c *ProductServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListProductResponse, error) {
	var out v11.ListProductResponse
	pattern := "/app/v1/mall/products"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationProductServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}