# backend/app/app/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Content Hub App API
    description: GoWind Content Hub App API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-shop
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-shop/blob/master/LICENSE
    version: "1.0"
paths:
...
    /app/v1/mall/comments:
        get:
            tags:
                - CommentService
            operationId: CommentService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/comment.service.v1.ListCommentResponse'
        post:
            tags:
                - CommentService
            operationId: CommentService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/comment.service.v1.CreateCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
        delete:
            tags:
                - CommentService
            operationId: CommentService_Delete
            parameters:
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
    /app/v1/mall/comments/{id}:
        get:
            tags:
                - CommentService
            operationId: CommentService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/comment.service.v1.Comment'
        put:
            tags:
                - CommentService
            operationId: CommentService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/comment.service.v1.UpdateCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
...
    /app/v1/mall/products:
        get:
            tags:
                - ProductService
            description: 分页查询商品列表
            operationId: ProductService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/catalog.service.v1.ListProductResponse'
    /app/v1/mall/products/{id}:
        get:
            tags:
                - ProductService
            description: 查询商品详情
            operationId: ProductService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: code
                  in: query
                  schema:
                    type: string
                - name: locale
                  in: query
                  schema:
                    type: string
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/catalog.service.v1.Product'
...
components:
    schemas:
...
        catalog.service.v1.ListProductResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/catalog.service.v1.Product'
                total:
                    type: string
            description: 商品列表 - 答复
...
        catalog.service.v1.Product:
            type: object
            properties:
                id:
                    type: integer
                    description: 商品ID
                    format: uint32
                status:
                    enum:
                        - PRODUCT_STATUS_UNSPECIFIED
                        - PRODUCT_STATUS_DRAFT
                        - PRODUCT_STATUS_ACTIVE
                        - PRODUCT_STATUS_INACTIVE
                    type: string
                    description: 商品状态
                    format: enum
                categoryId:
                    type: integer
                    description: 所属类目ID
                    format: uint32
                brandId:
                    type: integer
                    description: 所属品牌ID
                    format: uint32
                sortOrder:
                    type: integer
                    description: 排序优先级
                    format: uint32
                imageUrl:
                    type: string
                    description: 商品主图资源 URL（locale 无关）
                translations:
                    type: array
                    items:
                        $ref: '#/components/schemas/catalog.service.v1.ProductTranslation'
                    description: 多语言翻译列表
                availableLanguages:
                    example: ["zh-CN", "en-US"]
                    type: array
                    items:
                        type: string
                    description: 可用的语言代码列表
                createdBy:
                    type: integer
                    description: 创建者用户ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者用户ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 商品
...
        catalog.service.v1.ProductTranslation:
            type: object
            properties:
                id:
                    type: integer
                    description: 翻译记录ID
                    format: uint32
                productId:
                    type: integer
                    description: 关联的商品ID
                    format: uint32
                languageCode:
                    type: string
                    description: 语言代码
                name:
                    type: string
                    description: 商品名称
                slug:
                    type: string
                    description: 商品别名
                shortDescription:
                    type: string
                    description: 商品简短描述
                longDescription:
                    type: string
                    description: 商品长描述（HTML）
                createdBy:
                    type: integer
                    description: 创建者用户ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者用户ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 商品翻译
...
        comment.service.v1.Comment:
            type: object
            properties:
                id:
                    type: integer
                    description: 评论ID
                    format: uint32
                contentType:
                    enum:
                        - CONTENT_TYPE_UNSPECIFIED
                        - CONTENT_TYPE_PRODUCT
                    type: string
                    description: 内容类型
                    format: enum
                objectId:
                    type: integer
                    description: 对象ID
                    format: uint32
                content:
                    type: string
                    description: 评论内容
                status:
                    enum:
                        - STATUS_UNSPECIFIED
                        - STATUS_PENDING
                        - STATUS_APPROVED
                        - STATUS_REJECTED
                        - STATUS_SPAM
                    type: string
                    description: 评论状态
                    format: enum
                parentId:
                    type: integer
                    description: 父节点ID
                    format: uint32
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/comment.service.v1.Comment'
                    description: 子节点树
                tenantId:
                    type: integer
                    description: 租户ID
                    format: uint32
                createdBy:
                    type: integer
                    description: 创建者用户ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者用户ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 商品评论
        comment.service.v1.CreateCommentRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/comment.service.v1.Comment'
            description: 请求 - 创建评论
        comment.service.v1.ListCommentResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/comment.service.v1.Comment'
                total:
                    type: string
            description: 回应 - 评论列表
        comment.service.v1.UpdateCommentRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                data:
                    $ref: '#/components/schemas/comment.service.v1.Comment'
                updateMask:
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 允许缺失则插入
            description: 请求 - 更新评论
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...